+++ app/src/main/java/com/example/cook_book/RecipeModel.kt
@@ -1,18 +1,18 @@
 package com.example.cook_book
 
 import android.graphics.Bitmap
+import java.io.Serializable
 
-public data class Recipe(
+public data class Recipe (
     var recipeName: String,
     var recipeDescription: String,
     var imageName: String,
-    var image: Bitmap?,
     var ingredients: MutableList<String> = mutableListOf(),
     var instructions: MutableList<String> = mutableListOf(),
     var flipped: Boolean = false,
     var ingredientsSelect: Boolean = true
-)
+) : Serializable
 
-class RecipeModel : ArrayList<Recipe>() {
+class RecipeModel : ArrayList<Recipe>(), Serializable{
 
 }
